requirements.txt
# GUEZI RAG Chatbot Dependencies
# Rabbi Nachman/Breslov Texts with Gemini Flash 2.0
# Multi-language support with TTS

# Google Gemini AI
google-genai>=1.0.0

# Vector Database (FAISS)
faiss-cpu>=1.7.0
numpy>=1.24.0

# Sefaria API
requests>=2.31.0

# Web Interface
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
